Add SynchronizerIntegrationTests file
* Move External Id tests to SynchronizerIntegrationTests
