+++ OneSignalSDK/unittest/src/test/java/com/test/onesignal/MainOneSignalClassRunner.java
@@ -128,7 +128,6 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.regex.Pattern;
 
-import static com.onesignal.OneSignal.ExternalIdErrorType.REQUIRES_EXTERNAL_ID_AUTH;
 import static com.onesignal.OneSignalPackagePrivateHelper.FCMBroadcastReceiver_processBundle;
 import static com.onesignal.OneSignalPackagePrivateHelper.NotificationBundleProcessor_Process;
 import static com.onesignal.OneSignalPackagePrivateHelper.NotificationOpenedProcessor_processFromContext;
@@ -216,38 +215,6 @@ public class MainOneSignalClassRunner {
       };
    }
 
-   private static JSONObject lastExternalUserIdResponse;
-   private static OneSignal.ExternalIdError lastExternalUserIdError;
-   private static OneSignal.OSExternalUserIdUpdateCompletionHandler getExternalUserIdUpdateCompletionHandler() {
-      return new OneSignal.OSExternalUserIdUpdateCompletionHandler() {
-         @Override
-         public void onSuccess(JSONObject results) {
-            lastExternalUserIdResponse = results;
-         }
-
-         @Override
-         public void onFailure(OneSignal.ExternalIdError error) {
-            lastExternalUserIdError = error;
-         }
-      };
-   }
-
-   private static boolean didEmailUpdateSucceed;
-   private static OneSignal.EmailUpdateError lastEmailUpdateFailure;
-   private static OneSignal.EmailUpdateHandler getEmailUpdateHandler() {
-      return new OneSignal.EmailUpdateHandler() {
-         @Override
-         public void onSuccess() {
-            didEmailUpdateSucceed = true;
-         }
-
-         @Override
-         public void onFailure(OneSignal.EmailUpdateError error) {
-            lastEmailUpdateFailure = error;
-         }
-      };
-   }
-
    private static JSONObject lastGetTags;
    private static void getGetTagsHandler() {
       OneSignal.getTags(tags -> lastGetTags = tags);
@@ -256,10 +223,6 @@ public class MainOneSignalClassRunner {
    private static void cleanUp() throws Exception {
       lastNotificationOpenedBody = null;
       lastGetTags = null;
-      lastExternalUserIdResponse = null;
-      lastExternalUserIdError = null;
-      lastEmailUpdateFailure = null;
-      didEmailUpdateSucceed = false;
 
       ShadowGMSLocationController.reset();
 
@@ -4322,532 +4285,6 @@ public class MainOneSignalClassRunner {
       OneSignal.initWithContext(blankActivity);
       OneSignal.setNotificationOpenedHandler(getNotificationOpenedHandler());
       assertNull(ShadowFirebaseAnalytics.lastEventString);
-   }
-
-   @Test
-   public void shouldSendExternalUserIdAfterRegistration() throws Exception {
-      OneSignalInit();
-      threadAndTaskWait();
-
-      String testExternalId = "test_ext_id";
-
-      OneSignal.setExternalUserId(testExternalId);
-
-      threadAndTaskWait();
-
-      assertEquals(3, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request externalIdRequest = ShadowOneSignalRestClient.requests.get(2);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, externalIdRequest.method);
-      assertEquals(testExternalId, externalIdRequest.payload.getString("external_user_id"));
-   }
-
-   @Test
-   public void shouldSendExternalUserIdBeforeRegistration() throws Exception {
-      String testExternalId = "test_ext_id";
-
-      OneSignal.setExternalUserId(testExternalId);
-
-      OneSignalInit();
-      threadAndTaskWait();
-
-      assertEquals(2, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request registrationRequest = ShadowOneSignalRestClient.requests.get(1);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.POST, registrationRequest.method);
-      assertEquals(testExternalId, registrationRequest.payload.getString("external_user_id"));
-   }
-
-   @Test
-   public void shouldSetExternalIdWithAuthHash() throws Exception {
-      ShadowOneSignalRestClient.setRemoteParamsGetHtmlResponse(new JSONObject().put("require_user_id_auth", true));
-
-      OneSignalInit();
-      threadAndTaskWait();
-
-      String testExternalId = "test_ext_id";
-
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      assertNotNull(lastExternalUserIdError);
-      assertEquals(REQUIRES_EXTERNAL_ID_AUTH, lastExternalUserIdError.getType());
-   }
-
-   @Test
-   public void shouldSetExternalIdWithAuthHashAfterRegistration() throws Exception {
-      OneSignalInit();
-      threadAndTaskWait();
-
-      String testExternalId = "test_ext_id";
-      String mockExternalIdHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setExternalUserId(testExternalId, mockExternalIdHash, null);
-      threadAndTaskWait();
-
-      assertEquals(3, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request externalIdRequest = ShadowOneSignalRestClient.requests.get(2);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, externalIdRequest.method);
-      assertEquals(testExternalId, externalIdRequest.payload.getString("external_user_id"));
-      assertEquals(mockExternalIdHash, externalIdRequest.payload.getString("external_user_id_auth_hash"));
-   }
-
-   @Test
-   public void shouldSetExternalIdWithAuthHashBeforeRegistration() throws Exception {
-      String testExternalId = "test_ext_id";
-      String mockExternalIdHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setExternalUserId(testExternalId, mockExternalIdHash, null);
-
-      OneSignalInit();
-      threadAndTaskWait();
-
-      assertEquals(2, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request registrationRequest = ShadowOneSignalRestClient.requests.get(1);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.POST, registrationRequest.method);
-      assertEquals(testExternalId, registrationRequest.payload.getString("external_user_id"));
-      assertEquals(mockExternalIdHash, registrationRequest.payload.getString("external_user_id_auth_hash"));
-   }
-
-   @Test
-   public void shouldAlwaysSetExternalIdWithAuthHashAAfterRegistration() throws Exception {
-      OneSignalInit();
-      threadAndTaskWait();
-
-      String testExternalId = "test_ext_id";
-      String mockExternalIdHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setExternalUserId(testExternalId, mockExternalIdHash);
-      threadAndTaskWait();
-
-      ShadowOneSignalRestClient.Request registrationRequest = ShadowOneSignalRestClient.requests.get(2);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, registrationRequest.method);
-      assertEquals(testExternalId, registrationRequest.payload.getString("external_user_id"));
-      assertEquals(mockExternalIdHash, registrationRequest.payload.getString("external_user_id_auth_hash"));
-
-      fastColdRestartApp();
-
-      time.advanceSystemTimeBy(60);
-      OneSignalInit();
-      threadAndTaskWait();
-
-      ShadowOneSignalRestClient.Request registrationRequestAfterColdStart = ShadowOneSignalRestClient.requests.get(4);
-      assertEquals(REST_METHOD.POST, registrationRequestAfterColdStart.method);
-      assertEquals(mockExternalIdHash, registrationRequestAfterColdStart.payload.getString("external_user_id_auth_hash"));
-   }
-
-   @Test
-   public void shouldAlwaysSetExternalIdAndEmailWithAuthHashAfterRegistration() throws Exception {
-      OneSignalInit();
-      threadAndTaskWait();
-
-      String testExternalId = "test_ext_id";
-      String mockExternalIdHash = new String(new char[64]).replace('\0', '0');
-
-      String email = "josh@onesignal.com";
-      String mockEmailHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setExternalUserId(testExternalId, mockExternalIdHash);
-      OneSignal.setEmail(email, mockEmailHash);
-      threadAndTaskWait();
-
-      ShadowOneSignalRestClient.Request registrationRequest = ShadowOneSignalRestClient.requests.get(2);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, registrationRequest.method);
-      assertEquals(testExternalId, registrationRequest.payload.getString("external_user_id"));
-      assertEquals(mockExternalIdHash, registrationRequest.payload.getString("external_user_id_auth_hash"));
-
-      ShadowOneSignalRestClient.Request emailPost = ShadowOneSignalRestClient.requests.get(3);
-      assertEquals(REST_METHOD.POST, emailPost.method);
-      assertEquals(email, emailPost.payload.getString("identifier"));
-      assertEquals(11, emailPost.payload.getInt("device_type"));
-      assertEquals(mockEmailHash, emailPost.payload.getString("email_auth_hash"));
-
-      fastColdRestartApp();
-
-      time.advanceSystemTimeBy(60);
-      OneSignalInit();
-      threadAndTaskWait();
-
-      ShadowOneSignalRestClient.Request registrationRequestAfterColdStart = ShadowOneSignalRestClient.requests.get(6);
-      assertEquals(REST_METHOD.POST, registrationRequestAfterColdStart.method);
-      assertEquals(mockExternalIdHash, registrationRequestAfterColdStart.payload.getString("external_user_id_auth_hash"));
-
-      ShadowOneSignalRestClient.Request emailPostAfterColdStart = ShadowOneSignalRestClient.requests.get(7);
-      assertEquals(REST_METHOD.POST, emailPostAfterColdStart.method);
-      assertEquals(11, emailPostAfterColdStart.payload.getInt("device_type"));
-      assertEquals(mockEmailHash, emailPostAfterColdStart.payload.getString("email_auth_hash"));
-   }
-
-   @Test
-   public void shouldRemoveExternalUserId() throws Exception {
-      OneSignal.setExternalUserId("test_ext_id");
-
-      OneSignalInit();
-      threadAndTaskWait();
-
-      OneSignal.removeExternalUserId();
-      threadAndTaskWait();
-
-      assertEquals(3, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request removeIdRequest = ShadowOneSignalRestClient.requests.get(2);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, removeIdRequest.method);
-      assertEquals(removeIdRequest.payload.getString("external_user_id"), "");
-   }
-
-   @Test
-   public void shouldRemoveExternalUserIdFromPushWithAuthHash() throws Exception {
-      ShadowOneSignalRestClient.setRemoteParamsGetHtmlResponse(new JSONObject().put("require_user_id_auth", true));
-      
-      String testExternalId = "test_ext_id";
-      String mockExternalIdHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setExternalUserId(testExternalId, mockExternalIdHash, null);
-      OneSignalInit();
-      threadAndTaskWait();
-
-      OneSignal.removeExternalUserId(getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-                      "   \"push\" : {" +
-                      "      \"success\" : true" +
-                      "   }" +
-                      "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-
-      assertEquals(3, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request removeIdRequest = ShadowOneSignalRestClient.requests.get(2);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, removeIdRequest.method);
-      assertEquals(removeIdRequest.payload.getString("external_user_id"), "");
-      assertEquals(mockExternalIdHash, removeIdRequest.payload.getString("external_user_id_auth_hash"));
-   }
-
-   @Test
-   public void shouldRemoveExternalUserIdFromEmailWithAuthHash() throws Exception {
-      String testEmail = "test@test.com";
-      String mockEmailHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setEmail(testEmail, mockEmailHash, getEmailUpdateHandler());
-      OneSignalInit();
-      threadAndTaskWait();
-
-      OneSignal.removeExternalUserId(getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-                      "   \"push\" : {" +
-                      "      \"success\" : true" +
-                      "   }" + ", " +
-                      "   \"email\" : {" +
-                      "      \"success\" : true" +
-                      "   }" +
-                      "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-      assertTrue(didEmailUpdateSucceed);
-      assertNull(lastEmailUpdateFailure);
-
-      assertEquals(6, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request removeIdRequest = ShadowOneSignalRestClient.requests.get(4);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, removeIdRequest.method);
-      assertEquals(removeIdRequest.payload.getString("external_user_id"), "");
-      assertFalse(removeIdRequest.payload.has("external_user_id_auth_hash"));
-
-      ShadowOneSignalRestClient.Request removeIdEmailRequest = ShadowOneSignalRestClient.requests.get(5);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, removeIdEmailRequest.method);
-      assertEquals(removeIdEmailRequest.payload.getString("external_user_id"), "");
-      assertEquals(mockEmailHash, removeIdEmailRequest.payload.getString("email_auth_hash"));
-   }
-
-   @Test
-   public void shouldRemoveExternalUserIdFromPushAndEmailWithAuthHash() throws Exception {
-      String testExternalId = "test_ext_id";
-      String mockExternalIdHash = new String(new char[64]).replace('\0', '0');
-      String testEmail = "test@test.com";
-      String mockEmailHash = new String(new char[64]).replace('\0', '0');
-
-      OneSignal.setExternalUserId(testExternalId, mockExternalIdHash, null);
-      OneSignal.setEmail(testEmail, mockEmailHash, null);
-      OneSignalInit();
-      threadAndTaskWait();
-
-      OneSignal.removeExternalUserId();
-      threadAndTaskWait();
-
-      assertEquals(6, ShadowOneSignalRestClient.networkCallCount);
-
-      ShadowOneSignalRestClient.Request removeIdRequest = ShadowOneSignalRestClient.requests.get(4);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, removeIdRequest.method);
-      assertEquals(removeIdRequest.payload.getString("external_user_id"), "");
-      assertEquals(mockExternalIdHash, removeIdRequest.payload.getString("external_user_id_auth_hash"));
-
-      ShadowOneSignalRestClient.Request removeIdEmailRequest = ShadowOneSignalRestClient.requests.get(5);
-      assertEquals(ShadowOneSignalRestClient.REST_METHOD.PUT, removeIdEmailRequest.method);
-      assertEquals(removeIdEmailRequest.payload.getString("external_user_id"), "");
-      assertEquals(mockEmailHash, removeIdEmailRequest.payload.getString("email_auth_hash"));
-   }
-
-   @Test
-   public void doesNotSendSameExternalId() throws Exception {
-      String testExternalId = "test_ext_id";
-      OneSignal.setExternalUserId(testExternalId);
-
-      OneSignalInit();
-      threadAndTaskWait();
-
-      assertEquals(2, ShadowOneSignalRestClient.networkCallCount);
-
-      OneSignal.setExternalUserId(testExternalId);
-      threadAndTaskWait();
-
-      // Setting the same ID again should not generate a duplicate API request
-      // The SDK should detect it is the same and not generate a request
-      assertEquals(2, ShadowOneSignalRestClient.networkCallCount);
-   }
-
-   @Test
-   public void sendsExternalIdOnEmailPlayers() throws Exception {
-      String testExternalId = "test_ext_id";
-
-      OneSignalInit();
-      threadAndTaskWait();
-
-      OneSignal.setEmail("brad@onesignal.com");
-      threadAndTaskWait();
-
-      int currentRequestCount = ShadowOneSignalRestClient.networkCallCount;
-
-      OneSignal.setExternalUserId(testExternalId);
-      threadAndTaskWait();
-
-      // the SDK should have made two additional API calls
-      // One to set extID on the push player record,
-      // and another for the email player record
-      assertEquals(ShadowOneSignalRestClient.networkCallCount, currentRequestCount + 2);
-
-      int externalIdRequests = 0;
-
-      for (ShadowOneSignalRestClient.Request request : ShadowOneSignalRestClient.requests) {
-         if (request.payload != null && request.payload.has("external_user_id")) {
-            externalIdRequests += 1;
-            assertEquals(request.payload.getString("external_user_id"), testExternalId);
-         }
-      }
-
-      assertEquals(externalIdRequests, 2);
-   }
-
-   @Test
-   public void sendExternalUserId_withCompletionHandler() throws Exception {
-      String testExternalId = "test_ext_id";
-
-      // 1. Init OneSignal
-      OneSignalInit();
-      threadAndTaskWait();
-
-      // 2. Attempt to set external user id with callback
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 3. Make sure lastExternalUserIdResponse is equal to the expected response
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : true" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-   }
-
-    @Test
-    public void sendDifferentExternalUserId_withCompletionHandler() throws Exception {
-        String testExternalId = "test_ext_id_1";
-
-        // 1. Init OneSignal
-        OneSignalInit();
-        threadAndTaskWait();
-
-        // 2. Attempt to set external user id with callback
-        OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-        threadAndTaskWait();
-
-       // 3. Make sure lastExternalUserIdResponse is equal to the expected response
-       JSONObject expectedExternalUserIdResponse = new JSONObject(
-               "{" +
-               "   \"push\" : {" +
-               "      \"success\" : true" +
-               "   }" +
-               "}"
-       );
-       assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-
-        // 4. Change test external user id to send
-        testExternalId = "test_ext_id_2";
-
-        // 5. Attempt to set same exact external user id with callback
-        OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-        threadAndTaskWait();
-
-       // 6. Make sure lastExternalUserIdResponse is equal to the expected response
-       assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-    }
-
-   @Test
-   public void sendSameExternalUserId_withCompletionHandler() throws Exception {
-      String testExternalId = "test_ext_id";
-
-      // 1. Init OneSignal
-      OneSignalInit();
-      threadAndTaskWait();
-
-      // 2. Attempt to set external user id with callback
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 3. Make sure lastExternalUserIdResponse is equal to the expected response
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : true" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-
-      // 4. Attempt to set same exact external user id with callback
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 5. Make sure lastExternalUserIdResponse is equal to the expected response
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-   }
-
-   @Test
-   public void sendExternalUserId_withFailure_withCompletionHandler() throws Exception {
-      String testExternalId = "test_ext_id";
-
-      // 1. Init OneSignal
-      OneSignalInit();
-      threadAndTaskWait();
-
-      // 2. Attempt to set external user id with callback and force failure on the network requests
-      ShadowOneSignalRestClient.failAll = true;
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 3. Make sure lastExternalUserIdResponse is equal to the expected response
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : false" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-
-      // 4. Flip ShadowOneSignalRestClient.failAll flag back to false
-      ShadowOneSignalRestClient.failAll = false;
-
-      // 5. Attempt a second set external user id with callback
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 6. Make sure lastExternalUserIdResponse is equal to the expected response
-      expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : true" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-   }
-
-   @Test
-   public void sendExternalUserId_forPushAndEmail_withFailure_withCompletionHandler() throws Exception {
-      String testEmail = "test@onesignal.com";
-      String testExternalId = "test_ext_id";
-
-      // 1. Init OneSignal
-      OneSignalInit();
-      OneSignal.setEmail(testEmail);
-      threadAndTaskWait();
-
-      // 2. Attempt to set external user id with callback and force failure on the network requests
-      ShadowOneSignalRestClient.failAll = true;
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 3. Make sure lastExternalUserIdResponse has push and email with success : false
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : false" +
-              "   }," +
-              "   \"email\" : {" +
-              "      \"success\" : false" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-
-      // 4. Flip ShadowOneSignalRestClient.failAll flag back to false
-      ShadowOneSignalRestClient.failAll = false;
-
-      // 5. Attempt a second set external user id with callback
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 6. Make sure lastExternalUserIdResponse has push and email with success : true
-      expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : true" +
-              "   }," +
-              "   \"email\" : {" +
-              "      \"success\" : true" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
-   }
-
-   @Test
-   public void sendExternalUserId_forPush_afterLoggingOutEmail_withCompletion() throws Exception {
-      String testEmail = "test@onesignal.com";
-      String testExternalId = "test_ext_id";
-
-      // 1. Init OneSignal and set email
-      OneSignalInit();
-      OneSignal.setEmail(testEmail);
-      threadAndTaskWait();
-
-      // 2. Logout Email
-      OneSignal.logoutEmail();
-      threadAndTaskWait();
-
-      // 4. Attempt a set external user id with callback
-      OneSignal.setExternalUserId(testExternalId, getExternalUserIdUpdateCompletionHandler());
-      threadAndTaskWait();
-
-      // 5. Make sure lastExternalUserIdResponse has push with success : true
-      JSONObject expectedExternalUserIdResponse = new JSONObject(
-              "{" +
-              "   \"push\" : {" +
-              "      \"success\" : true" +
-              "   }" +
-              "}"
-      );
-      assertEquals(expectedExternalUserIdResponse.toString(), lastExternalUserIdResponse.toString());
    }
 
    @Test
